cleaned up an make stuff work
